new: Todo list will still show while loading
Improves loading indicator so that the Todo list will still show while
loading is happening (assuming the list has already been received at
least once).

This does cause a slight sync confusion where when the list is
re-ordered with drag and drop the user will see the new list before the
server has updated. It still informs the user that it is loading, but
this messaging could be improved.